Lights, camera, background, player movement, shots
Created lighting and set up the camera for the game. Added a background.
Added a player controller script to the player to allow movement and
shooting. Created a bolt prefab for the player's shots and added a
boundary that can destroy shots as they leave the game view. Added a
mover script and destroy by boundary script to handle moving and
destroying shots.