Pooling, bolt lighting, randomized enemy shot sound
Changes in this commit onward are personal upgrades to the project that
were not included in the original tutorial. Created orange and cyan
lighting effects for the player and enemy shots. Made the enemy shot
sound have slightly randomized volume and pitch. Added premade mobile
art assets. Implemented an object pooler which allows object pooling for
asteroids and shots.